GetUserIdmResourceTest test is extracted
