+++ src/test/java/gov/ca/cwds/idm/FirstIdmResourceTest.java
@@ -3,19 +3,10 @@ package gov.ca.cwds.idm;
 import static gov.ca.cwds.config.api.idm.Roles.CALS_ADMIN;
 import static gov.ca.cwds.config.api.idm.Roles.OFFICE_ADMIN;
 import static gov.ca.cwds.config.api.idm.Roles.STATE_ADMIN;
-import static gov.ca.cwds.idm.TestCognitoServiceFacade.ABSENT_USER_ID;
-import static gov.ca.cwds.idm.TestCognitoServiceFacade.COUNTY_ADMIN_ID;
-import static gov.ca.cwds.idm.TestCognitoServiceFacade.ERROR_USER_ID;
 import static gov.ca.cwds.idm.TestCognitoServiceFacade.ES_ERROR_CREATE_USER_EMAIL;
 import static gov.ca.cwds.idm.TestCognitoServiceFacade.NEW_USER_ES_FAIL_ID;
 import static gov.ca.cwds.idm.TestCognitoServiceFacade.NEW_USER_SUCCESS_ID;
 import static gov.ca.cwds.idm.TestCognitoServiceFacade.SOME_PAGINATION_TOKEN;
-import static gov.ca.cwds.idm.TestCognitoServiceFacade.STATE_ADMIN_ID;
-import static gov.ca.cwds.idm.TestCognitoServiceFacade.USER_CALS_EXTERNAL;
-import static gov.ca.cwds.idm.TestCognitoServiceFacade.USER_NO_RACFID_ID;
-import static gov.ca.cwds.idm.TestCognitoServiceFacade.USER_WITH_NO_PHONE_EXTENSION;
-import static gov.ca.cwds.idm.TestCognitoServiceFacade.USER_WITH_RACFID_AND_DB_DATA_ID;
-import static gov.ca.cwds.idm.TestCognitoServiceFacade.USER_WITH_RACFID_ID;
 import static gov.ca.cwds.idm.util.AssertFixtureUtils.assertExtensible;
 import static gov.ca.cwds.idm.util.AssertFixtureUtils.assertNonStrict;
 import static org.hamcrest.CoreMatchers.is;
@@ -41,108 +32,6 @@ import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 
 public class FirstIdmResourceTest extends IdmResourceTest {
-
-  @Test
-  @WithMockCustomUser
-  public void testGetUserNoRacfId() throws Exception {
-    testGetValidUser(USER_NO_RACFID_ID, "fixtures/idm/get-user/no-racfid-valid.json");
-  }
-
-  @Test
-  @WithMockCustomUser(roles = {OFFICE_ADMIN})
-  public void testGetUserOfficeAdmin() throws Exception {
-    testGetValidUser(USER_WITH_RACFID_AND_DB_DATA_ID,
-        "fixtures/idm/get-user/with-racfid-and-db-data-valid-3.json");
-  }
-
-  @Test
-  @WithMockCustomUser(roles = {OFFICE_ADMIN}, adminOfficeIds = {"otherOfficeId"})
-  public void testGetUserOfficeAdminOtherOffice() throws Exception {
-    testGetValidUser(USER_WITH_RACFID_AND_DB_DATA_ID,
-        "fixtures/idm/get-user/with-racfid-and-db-data-valid-2.json");
-  }
-
-  @Test
-  @WithMockCustomUser(roles = {CALS_ADMIN})
-  public void testGetUserCalsAdminUnauthorized() throws Exception {
-    assertGetUserUnauthorized(USER_NO_RACFID_ID);
-  }
-
-  @Test
-  @WithMockCustomUser(roles = {CALS_ADMIN})
-  public void testGetUserCalsAdmin() throws Exception {
-    testGetValidUser(USER_CALS_EXTERNAL,
-        "fixtures/idm/get-user/with-cals-externa-worker-role.json");
-  }
-
-  @Test
-  @WithMockCustomUser
-  public void testGetUserWithRacfId() throws Exception {
-    testGetValidUser(USER_WITH_RACFID_ID, "fixtures/idm/get-user/with-racfid-valid-1.json");
-  }
-
-  @Test
-  @WithMockCustomUser
-  public void testGetUserWithRacfIdAndDbData() throws Exception {
-    testGetValidUser(
-        USER_WITH_RACFID_AND_DB_DATA_ID,
-        "fixtures/idm/get-user/with-racfid-and-db-data-valid-1.json");
-  }
-
-  @Test
-  @WithMockCustomUser
-  public void testGetUserWithNoPhoneExtension() throws Exception {
-    testGetValidUser(
-        USER_WITH_NO_PHONE_EXTENSION,
-        "fixtures/idm/get-user/with-racfid-and-no-phone-extension.json");
-  }
-
-  @Test
-  @WithMockCustomUser
-  public void testGetAbsentUser() throws Exception {
-
-    mockMvc
-        .perform(MockMvcRequestBuilders.get("/idm/users/" + ABSENT_USER_ID))
-        .andExpect(MockMvcResultMatchers.status().isNotFound())
-        .andReturn();
-  }
-
-  @Test
-  @WithMockCustomUser
-  public void testGetUserError() throws Exception {
-    assertGetUserUnauthorized(ERROR_USER_ID);
-  }
-
-  @Test
-  @WithMockCustomUser(county = "Madera")
-  public void testGetUserByOtherCountyAdmin() throws Exception {
-    assertGetUserUnauthorized(USER_NO_RACFID_ID);
-  }
-
-  @Test
-  @WithMockCustomUser(roles = {"OtherRole"})
-  public void testGetUserWithOtherRole() throws Exception {
-    assertGetUserUnauthorized(USER_NO_RACFID_ID);
-  }
-
-  @Test
-  @WithMockCustomUser(roles = {OFFICE_ADMIN}, adminOfficeIds = {"OtherOfficeId"})
-  public void testGetOtherOfficeCountyAdminByOfficeAdmin() throws Exception {
-    assertGetUserUnauthorized(COUNTY_ADMIN_ID);
-  }
-
-  @Test
-  @WithMockCustomUser(roles = {OFFICE_ADMIN}, adminOfficeIds = {"OtherOfficeId"})
-  public void testGetOtherOfficeStateAdminByOfficeAdmin() throws Exception {
-    assertGetUserUnauthorized(STATE_ADMIN_ID);
-  }
-
-  @Test
-  @WithMockCustomUser(roles = {STATE_ADMIN}, county = "Madera")
-  public void testGetUserStateAdminDifferentCounty() throws Exception {
-    testGetValidUser(USER_WITH_RACFID_ID, "fixtures/idm/get-user/with-racfid-valid-2.json");
-  }
-
 
   @Test
   public void testGetUsers() throws Exception {
